move add_label to post processing helpers
