+++ anomalib/post_processing/__init__.py
@@ -15,10 +15,19 @@
 # and limitations under the License.
 
 from .post_process import (
+    add_anomalous_label,
+    add_normal_label,
     anomaly_map_to_color_map,
     compute_mask,
     superimpose_anomaly_map,
 )
 from .visualizer import Visualizer
 
-__all__ = ["anomaly_map_to_color_map", "superimpose_anomaly_map", "compute_mask", "Visualizer"]
+__all__ = [
+    "add_anomalous_label",
+    "add_normal_label",
+    "anomaly_map_to_color_map",
+    "superimpose_anomaly_map",
+    "compute_mask",
+    "Visualizer",
+]
