+++ anomalib/post_processing/post_process.py
@@ -15,9 +15,73 @@
 # and limitations under the License.
 
 
+import math
+from typing import Optional, Tuple
+
 import cv2
 import numpy as np
 from skimage import morphology
+
+
+def add_label(
+    image: np.ndarray,
+    label_name: str,
+    color: Tuple[int, int, int],
+    confidence: Optional[float] = None,
+    font_scale: float = 5e-3,
+    thickness_scale=1e-3,
+):
+    """Adds a label to an image.
+
+    Args:
+        image (np.ndarray): Input image.
+        label_name (str): Name of the label that will be displayed on the image.
+        color (Tuple[int, int, int]): RGB values for background color of label.
+        confidence (Optional[float]): confidence score of the label.
+        font_scale (float): scale of the font size relative to image size. Increase for bigger font.
+        thickness_scale (float): scale of the font thickness. Increase for thicker font.
+
+    Returns:
+        np.ndarray: Image with label.
+    """
+    image = image.copy()
+    img_height, img_width, _ = image.shape
+
+    font = cv2.FONT_HERSHEY_PLAIN
+    text = label_name if confidence is None else f"{label_name} ({confidence*100:.0f}%)"
+
+    # get font sizing
+    font_scale = min(img_width, img_height) * font_scale
+    thickness = math.ceil(min(img_width, img_height) * thickness_scale)
+    (width, height), baseline = cv2.getTextSize(text, font, fontScale=font_scale, thickness=thickness)
+
+    # create label
+    label_patch = np.zeros((height + baseline, width + baseline, 3), dtype=np.uint8)
+    label_patch[:, :] = color
+    cv2.putText(
+        label_patch,
+        text,
+        (0, baseline // 2 + height),
+        font,
+        fontScale=font_scale,
+        thickness=thickness,
+        color=0,
+        lineType=cv2.LINE_AA,
+    )
+
+    # add label to image
+    image[: baseline + height, : baseline + width] = label_patch
+    return image
+
+
+def add_normal_label(image: np.ndarray, confidence: Optional[float] = None):
+    """Adds the normal label to the image."""
+    return add_label(image, "normal", (225, 252, 134), confidence)
+
+
+def add_anomalous_label(image: np.ndarray, confidence: Optional[float] = None):
+    """Adds the anomalous label to the image."""
+    return add_label(image, "anomalous", (255, 100, 100), confidence)
 
 
 def anomaly_map_to_color_map(anomaly_map: np.ndarray, normalize: bool = True) -> np.ndarray:
